wince: probe late etimer helper and test replay consume hook
Tried:
- fixed the replay PC probe bookkeeping so late WinCE probe entries beyond 32 labels do not alias and disappear
- added late-handler WinCE replay probes for 0x800AAD2C, 0x800AAE20, and 0x800AAE6C in both the generic replay logger and the VR41xx MMIO-triggered probe path
- added a replay-only ETIMER consume experiment that clears pending timer state when WinCE executes the 0x800AAE3C helper write to PMUCNTREG

Outcome:
- decoded the live late helper successfully: 0x800AAE20..0x800AAE40 only clears GIUINTENL bit 1 and PMUCNTREG bit 13, and 0x800AADA0 is the caller wrapper that invokes it
- the replay-only ETIMER consume hook fires repeatedly and proves the pending queue can be collapsed at that point, but WinCE still returns to the SYSINT1 bit 3 path and falls back into the late loop
- Linux regression still reaches mounted root and BusyBox init/userspace

Learned:
- the remaining blocker is upstream of the simple pending timer queue; a real interrupt source, registration path, or callback object is still reasserting the ETIMER path after the helper runs
- the late helper is a bit-clear wrapper, not the root cause by itself

Next step:
- trace the caller and registration path feeding 0x800AADA0 and the follow-on indirect path around 0x800AAE48 and 0x80660328 to determine whether replay restored the wrong callback object or is missing the state that should switch handlers

+++ src/wince_boot_types.h
@@ -92,7 +92,7 @@ typedef struct {
     uint32_t replay_region_drift_logged_mask;
     uint32_t replay_region_write_logged_mask;
     uint32_t replay_region_mismatch_logged_mask;
-    uint32_t replay_pc_probe_logged_mask;
+    uint64_t replay_pc_probe_logged_mask;
     uint32_t replay_bootctx_epc_probe_logged_mask;
     uint32_t replay_watch_write_logged_mask;
     uint32_t replay_mmio_watch_logged_mask;
